Remove utils, move its methods into the dropdown browser view and eliminate the need to call a module from the page template.
